Rename AnalogAcquisitionHandler to new oscilloscope.Oscilloscope
